Refactor runtime into modular structure
Split the monolithic runtime/builtins.c (4594 lines) into three focused modules:

- runtime/gc.c (835 lines) — Garbage collector with mark & sweep, STW coordination
- runtime/threading.c (2049 lines) — Concurrency primitives (tasks, channels, select)
- runtime/builtins.c (1786 lines) — Core runtime utilities (strings, arrays, I/O, maps)
- runtime/builtins.h (149 lines) — Shared declarations and GC tags

Module dependencies: threading.c and builtins.c depend on gc.c for allocation.
GC has no dependencies and serves as the foundational layer.

Key changes:
- Updated build system to compile 3 C files separately and link with ld -r
- Fixed threading.c to use GC API instead of direct internal access
- Added __pluto_gc_bytes_allocated() to test mode for heap size queries
- Created comprehensive runtime architecture documentation
- Updated CLAUDE.md and concurrency.md to reflect new structure

All 499 lib tests and 530+ integration tests pass in both test and production modes.
